Replace adjustPyClass with adjustInterfaceIR
For adding the extra signature to the json method, this is easier since we can
operate on a richer data structure than a text blob.

For deciding the format of the numbers, we can't use replaceAll anymore. Instead
I added a separate type to the IR to indicate that we have the conversion of
number, and added an extra setting to the type conversion to indicate how
number should be rendered. By default it is `int | float`, but we can modify it
to `int`.

Since the number of settings for the type to IR converter was getting out of
hand, I added a settings object for it. Probably we should add these for the
other conversion functions too...