Don't create bad types in type literal substituion in destructured arguments (#71)
In `getPropType()` we sometimes notice that there is a type param substitution so that the text of the
type of the prop node does not match the text of its definition node. In this case, we were calling
`typeToIR` twice and throwing away the first outcome. However, now `typeToIR()` can cause
side effects so we should make sure to only call it when we need the outcome.